WIP: A remote client for SparkContext.
This is a p.o.c. of a simple remote client interface for SparkContext. It does
not aim at providing remote equivalents of all of the SparkContext API. Instead,
it focuses on providing basic configuration APIs for the SparkContext and a
generic "submit job" interface that can be used to submit any code to be run on
the remote process managing the SparkContext.