Send metrics back to the client instance.
This is a little hackish, since it requires some communication between
the driver code and the job context so that jobs triggered by the user
can be monitored. It also requires user opt-in - right now an explicit
method call, although it could be changed to use a custom AsyncRDDActions
implementation instead (which would probably lead to more maintenance
in the long run).

There are also some question marks and some missing APIs in Spark to
be able to properly support this. But it's a start.

Finally, there is an open question about the safety of the current
RPC implementation: akka is asynchronous, although it says it does
guarantee messages arrive in order (which is why the code goes through
some trouble to make sure job results are sent only after all Spark
events related to the job are processed). But what about processing
the messages on the receiving side? Are they guaranteed to be processed
in order (by a single thread)?