Prefer spark-submit over SparkSubmit.
If "spark.home" is set, use the spark-submit script to start the
Spark job. This solves some issues that arise when using SparkSubmit
directly, such as needing to provide the right Guava version or other
needed dependencies that are not part of spark-core (for example, the
Yarn backend, or Streaming classes, or...).